Namespace, y u no work?
Needed to add a separate login and logout file to bypass wordpress authentication before headers are sent (necessary for raven to work properly).

Proof of concept works, just need to add support for redirection to the page the user came from to the login page.

Not sure why the namespace doesn't work, but when login.php, raven.php and ucam_webauth.php are namespaced, the constructor for ucam_webauth doesn't appear to be called.